restructured header file location
